Configure CAPA controller role to allow Crossplane to use it
This grants access to the Crossplane providers running on the MC to assume the CAPA controller role to perform actions on the target AWS account.

It also adds the necessary permissions to create the needed resources for Karpenter.

Since Crossplane uses IRSA, the installation OIDC domain is needed to configure the role trust policy.