Fix the MC oidc provider variable definition
The "_domain" suffix wasn't really accurate, since the oidc name can be in a URL form like `s3.cn-north-1.amazonaws.com.cn/some-s3-bucket`, which we actually use for some China MCs.

I've fixed the validation Regex to accept that case and renamed the variable to a more suitable name.